requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.0
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
replicate==0.28.0
lob==6.0.0
qrcode[pil]==7.4.2
python-multipart==0.0.9
websockets==12.0
aiofiles==23.2.1
jinja2==3.1.4
